Added the unit tests for Bernoulli Naive Bayes, obviously all fails for now.
Added the implementation for BernoulliNB, but the tests that uses binarization are failing as it is not supported in the python lib. The BernoulliNB remains to be finalized.

Finalized BernoulliNB. All tests passed.

Adapted the new changes from the python library:
* Load the data from external files instead of the binary package.
* Use the feature count & name in the base class.

Made the modification pass the checks of check_java_source.py.
Incremented the version to 0.0.3.

Added implementation of categorical naive bayes and tests (all passing).

Added implementation of complement naive bayes and tests (all passing).

Added implementation of multinomial naive bayes and tests (all passing).